[prototype runtime] Generate block argument with `yield` by RubyVM::AST.of
