+++ lib/rbs.rb
@@ -32,6 +32,7 @@ require "rbs/constant"
 require "rbs/constant_table"
 require "rbs/ast/comment"
 require "rbs/writer"
+require "rbs/prototype/helpers"
 require "rbs/prototype/rbi"
 require "rbs/prototype/rb"
 require "rbs/prototype/runtime"
